Struct fedimint_ln_server::Lightning

pub struct Lightning {
    pub(crate) cfg: LightningConfig,
    pub(crate) btc_rpc: DynBitcoindRpc,
}

The lightning module implements an account system. It does not have the privacy guarantees of the e-cash mint module but instead allows for smart contracting. There exist two contract types that can be used to “lock” accounts:

Outgoing: an account locked with an HTLC-like contract allowing to incentivize an external Lightning node to make payments for the funder

Incoming: a contract type that represents the acquisition of a preimage belonging to a hash. Every incoming contract is preceded by an offer that specifies how much the seller is asking for the preimage to a particular hash. It also contains some threshold-encrypted data. Once the contract is funded the data is decrypted. If it is a valid preimage the contract’s funds are now accessible to the creator of the offer, if not they are accessible to the funder.

These two primitives allow to integrate the federation with the wider Lightning network through a centralized but untrusted (except for availability) Lightning gateway server.

Fields
cfg: LightningConfigbtc_rpc: DynBitcoindRpc
Implementations

impl Lightning

pub(crate) fn new( cfg: LightningConfig, task_group: &mut TaskGroup ) -> Result<Self>

pub(crate) async fn block_count(&self) -> Result<u64>

pub(crate) async fn consensus_block_count( &self, dbtx: &mut DatabaseTransactionRef<'_> ) -> u64

pub(crate) async fn wait_block_height( &self, block_height: u64, dbtx: &mut DatabaseTransactionRef<'_> )

pub(crate) fn validate_decryption_share( &self, peer: PeerId, share: &PreimageDecryptionShare, message: &EncryptedPreimage ) -> bool

pub(crate) async fn get_offer( &self, dbtx: &mut DatabaseTransactionRef<'_>, payment_hash: Hash ) -> Option<IncomingContractOffer>

pub(crate) async fn wait_offer( &self, context: &mut ApiEndpointContext<'_>, payment_hash: Hash ) -> IncomingContractOffer

pub(crate) async fn get_contract_account( &self, dbtx: &mut DatabaseTransactionRef<'_>, contract_id: ContractId ) -> Option<ContractAccount>

pub(crate) async fn wait_contract_account( &self, context: &mut ApiEndpointContext<'_>, contract_id: ContractId ) -> ContractAccount

pub(crate) async fn wait_outgoing_contract_account_cancelled( &self, context: &mut ApiEndpointContext<'_>, contract_id: ContractId ) -> ContractAccount

pub(crate) async fn wait_preimage_decrypted( &self, context: &mut ApiEndpointContext<'_>, contract_id: ContractId ) -> (IncomingContractAccount, Option<Preimage>)

pub(crate) fn get_incoming_contract_account( contract: ContractAccount ) -> IncomingContractAccount

pub(crate) async fn list_gateways( &self, dbtx: &mut DatabaseTransactionRef<'_> ) -> Vec<LightningGatewayAnnouncement>

pub(crate) async fn register_gateway( &self, dbtx: &mut DatabaseTransactionRef<'_>, gateway: LightningGatewayAnnouncement )

pub(crate) async fn delete_expired_gateways( &self, dbtx: &mut DatabaseTransactionRef<'_> )