Module fedimint_ln_server::db

Structs

AgreedDecryptionShareContractIdPrefix

AgreedDecryptionShareKey

Preimage decryption shares we received

AgreedDecryptionShareKeyPrefix

Preimage decryption shares we received

BlockCountVoteKey

BlockCountVotePrefix

ContractKey

ContractKeyPrefix

ContractUpdateKey

ContractUpdateKeyPrefix

DbKeyPrefixIter

An iterator over the variants of DbKeyPrefix

EncryptedPreimageIndexKey

We save the hash of the encrypted preimage from each accepted offer so that we can make sure that no preimage is used twice.

EncryptedPreimageIndexKeyPrefix

LightningAuditItemKeyPrefix

LightningGatewayKey

LightningGatewayKeyPrefix

OfferKey

OfferKeyPrefix

ProposeDecryptionShareKey

ProposeDecryptionShareKeyPrefix

Our preimage decryption shares that still need to be broadcasted

Enums

DbKeyPrefix

LightningAuditItemKey

We keep a separate mapping of incoming and outgoing ContractIds to Amounts, which allows us to quickly audit the total liabilities in the Lightning module.